PainGenerator_WebUI/requirements.txt
Flask
Pillow
nltk
numpy